requirements.txt
discord.py==1.3.1
pymongo==3.10.1
dnspython==1.16.0
requests==2.23.0
